added filter for bogus 0-values after reconnect
With version U1.2.6 of the ComfoConnect LAN C bridge, the bridge
started sending `0` for every monitored sensor once, before
sending the correct values.

To filter these values, we keep a set of sensor IDs for which
we received a `0`. If a given ID is not in the set when receiving
a `0`, the value is ignored. The set is replaced with a new empty
set on every re-connect.

fixes #44